Move AbstractMethodInvocationHandler into 'api' package
